refactor: segment engine simplification round, both runtimes
Nine verified simplifications from a measured study (each proven
equivalent by worktree prototypes, differential sweeps, and the full
nets; perf A/B'd against A/A noise controls):

- one memoized ε-closure with a block flavor replaces the hand-rolled
  dot-blocked DFS, and the blocked entry set now also drops the
  dot-protected consumers, deleting the per-state dot_protected bools
  (Rust) and the dps array (JS); wild_led becomes exact
- segmentize threads two state pieces instead of four: InSegment is
  just a nonempty buffer, and the open-absorber flags fold into a
  Boundary::Open variant handled once at the loop top
- SegIter drops its done flag (pos > len is already the sentinel), and
  the JS single-globstar heads drop headExhausted the same way
- ElemSeq.single_g / g_count narrow to u8 (128 -> 112 B per fork)
- JS deletes five byte-array copies of literals it already holds as
  latin1 strings; byte mode reads the strings via charCodeAt
- JS needsAsciiSeg derives from the ops in the builder, so star bodies
  no longer force the string path to bail to byte mode
- op_is_crossing_alt delegates to op_crosses_segment
- the JS fork-expansion driver moves into compile.js as compileSeqs,
  restoring file-by-file twin alignment (exports 5 -> 2)

Net -114 lines. Rust matcher memory -3..5%, JS matcher memory -20..30%
on the published corpus; match/compile timings inside the noise floor.
The parked goto-mask idea was prototyped, benched flat-to-slower, and
is rejected for good.

+++ crates/globstar/src/engine/segment/mod.rs
@@ -136,11 +136,13 @@ struct ElemSeq {
     elems: Box<[Elem]>,
 
     /// Index of the single globstar element when the sequence has
-    /// exactly one (fast anchored path); `usize::MAX` otherwise.
-    single_g: usize,
+    /// exactly one (fast anchored path); `u8::MAX` otherwise. Every
+    /// element owns ≥ 1 state and `MAX_SEQ_STATES` caps a sequence at
+    /// 63 elements, so an index always fits.
+    single_g: u8,
 
     /// Number of globstar elements.
-    g_count: usize,
+    g_count: u8,
 
     /// Single-globstar fast path: when every head element is a `Lit`,
     /// the joined head bytes (`"src/"` — each segment plus one `/`),
